requirements.txt
Flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
PyJWT==2.8.0
cryptography==41.0.7
pycryptodome==3.19.0
requests==2.31.0
werkzeug==3.0.1
Jinja2==3.1.2
click==8.1.7
itsdangerous==2.1.2
MarkupSafe==2.1.3
blinker==1.7.0
psycopg2-binary==2.9.11
gunicorn==21.2.0
pyotp==2.9.0
qrcode==7.4.2
Pillow>=10.2.0
# Zero Trust VPN Dependencies
# (Cryptography is already included above)